give result its own crate
