+++ harness/src/fuzz/mod.rs
@@ -4,11 +4,8 @@ pub mod firedancer;
 pub mod mollusk;
 
 use {
-    crate::{result::InstructionResult, Mollusk},
-    mollusk_svm_fuzz_fs::FsHandler,
-    solana_account::Account,
-    solana_instruction::Instruction,
-    solana_pubkey::Pubkey,
+    crate::Mollusk, mollusk_svm_fuzz_fs::FsHandler, mollusk_svm_result::result::InstructionResult,
+    solana_account::Account, solana_instruction::Instruction, solana_pubkey::Pubkey,
 };
 
 pub fn generate_fixtures_from_mollusk_test(
