+++ result/src/lib.rs
@@ -1,0 +1,3 @@
+//! SVM program execution results and validation.
+
+pub mod result;
