+++ result/src/result.rs
@@ -259,7 +259,7 @@ impl InstructionResult {
         pass
     }
 
-    pub(crate) fn absorb(&mut self, other: Self) {
+    pub fn absorb(&mut self, other: Self) {
         self.compute_units_consumed += other.compute_units_consumed;
         self.execution_time += other.execution_time;
         self.program_result = other.program_result;
@@ -753,5 +753,60 @@ impl Compare {
             Self::ReturnData,
             Self::all_resulting_accounts(),
         ]
+    }
+}
+
+#[cfg(feature = "fuzz")]
+impl From<&InstructionResult> for mollusk_svm_fuzz_fixture::effects::Effects {
+    fn from(input: &InstructionResult) -> Self {
+        let compute_units_consumed = input.compute_units_consumed;
+        let execution_time = input.execution_time;
+        let return_data = input.return_data.clone();
+
+        let program_result = match &input.program_result {
+            ProgramResult::Success => 0,
+            ProgramResult::Failure(e) => u64::from(e.clone()),
+            ProgramResult::UnknownError(_) => u64::MAX, //TODO
+        };
+
+        let resulting_accounts = input.resulting_accounts.clone();
+
+        Self {
+            compute_units_consumed,
+            execution_time,
+            program_result,
+            return_data,
+            resulting_accounts,
+        }
+    }
+}
+
+#[cfg(feature = "fuzz")]
+impl From<&mollusk_svm_fuzz_fixture::effects::Effects> for InstructionResult {
+    fn from(input: &mollusk_svm_fuzz_fixture::effects::Effects) -> Self {
+        use solana_instruction::error::InstructionError;
+
+        let compute_units_consumed = input.compute_units_consumed;
+        let execution_time = input.execution_time;
+        let return_data = input.return_data.clone();
+
+        let raw_result = if input.program_result == 0 {
+            Ok(())
+        } else {
+            Err(InstructionError::from(input.program_result))
+        };
+
+        let program_result = raw_result.clone().into();
+
+        let resulting_accounts = input.resulting_accounts.clone();
+
+        Self {
+            compute_units_consumed,
+            execution_time,
+            program_result,
+            raw_result,
+            return_data,
+            resulting_accounts,
+        }
     }
 }
